Use python guidelines PEP8 for syntax
It's good to use PEP8 guidelines since the indentation etc. is defined
in an acceptable and understable way. You can read the reasons for every
decision they made and they are ok.

I recommend you to use the syntastic plugin for vim with flake8 syntax
checker. It warns you with a sign on the line where an error or
warning occured.
An error could be something like python syntax errors. A warning could be
an unused module, variable or a indentation error etc.
It helps you to keep clean and reusable code!